Failed to switch meshlet to uint8, something in the shader get corrupted

+++ src/Application/resourceLoading.cpp
@@ -138,8 +138,8 @@ namespace BlitzenEngine
         {
             BlitML::Meshlet meshlet = {};
 
-            BlitCL::DynamicArray<uint32_t> meshletVertices(resources.vertices.GetSize() - previousVertexBufferSize);
-            meshletVertices.Fill(UINT32_MAX);
+            BlitCL::DynamicArray<uint8_t> meshletVertices(resources.vertices.GetSize() - previousVertexBufferSize);
+            meshletVertices.Fill(UINT8_MAX);
 
             newSurface.firstMeshlet = resources.meshlets.GetSize();
 
@@ -149,34 +149,34 @@ namespace BlitzenEngine
                 uint32_t vtxIndexB = resources.indices[i + 1];
                 uint32_t vtxIndexC = resources.indices[i + 2];
 
-                uint32_t& vtxA = meshletVertices[vtxIndexA];
-                uint32_t& vtxB = meshletVertices[vtxIndexB];
-                uint32_t& vtxC = meshletVertices[vtxIndexC];
+                uint8_t& vtxA = meshletVertices[vtxIndexA];
+                uint8_t& vtxB = meshletVertices[vtxIndexB];
+                uint8_t& vtxC = meshletVertices[vtxIndexC];
 
                 // If the current meshlet's vertex count + the vertices that are going to be added next is over the meshlet vertex limit, 
                 // It gets added to the dynamic array and a new entry is created
-                if (meshlet.vertexCount + (vtxA == UINT32_MAX) + (vtxB == UINT32_MAX) + (vtxC == UINT32_MAX) > 64 || meshlet.triangleCount >= 126)
+                if (meshlet.vertexCount + (vtxA == UINT8_MAX) + (vtxB == UINT8_MAX) + (vtxC == UINT8_MAX) > 64 || meshlet.triangleCount >= 126)
 		        {
                     newSurface.meshletCount++;
 			        resources.meshlets.PushBack(meshlet);
 
 			        for (size_t j = 0; j < meshlet.vertexCount; ++j)
-				        meshletVertices[meshlet.vertices[j]] = UINT32_MAX;
+				        meshletVertices[meshlet.vertices[j]] = UINT8_MAX;
 			        
                     meshlet = {};
 		        }
 
-                if (vtxA == UINT32_MAX)
+                if (vtxA == UINT8_MAX)
 		        {
 		        	vtxA = meshlet.vertexCount;
 		        	meshlet.vertices[meshlet.vertexCount++] = vtxIndexA;
 		        }
-		        if (vtxB == UINT32_MAX)
+		        if (vtxB == UINT8_MAX)
 		        {
 		        	vtxB = meshlet.vertexCount;
 		        	meshlet.vertices[meshlet.vertexCount++] = vtxIndexB;
 		        }
-		        if (vtxC == UINT32_MAX)
+		        if (vtxC == UINT8_MAX)
 		        {
 		        	vtxC = meshlet.vertexCount;
 		        	meshlet.vertices[meshlet.vertexCount++] = vtxIndexC;
